feat: form fields
Created form sections with multiple inputs. We will start with the simplest inputs, such as the name input and the email input. We will use CSS Grid to organize the elements and define the size of the columns, we have a basic structure of the form ready to add the other inputs